feat: DataFrameCache using binary HDF5 format
HDF5 maintains indices and datetime information
