+++ src/workflowlib/__init__.py
@@ -2,5 +2,6 @@
 __version__ = '0.1.4'
 
 from . import dataframes, loaders, serializers
+from .process import DelegatedSource
 from .registry import register
 from .workflow import Workflow, run
